+++ src/utils.ts
@@ -32,9 +32,6 @@ export default class Utils {
       lightWorkbench: workspaceConfiguration.get<string>("lightWorkbench"),
       italicKeywords: workspaceConfiguration.get<boolean>("italicKeywords"),
       italicComments: workspaceConfiguration.get<boolean>("italicComments"),
-      diagnosticTextBackgroundOpacity: workspaceConfiguration.get<string>(
-        "diagnosticTextBackgroundOpacity"
-      ),
       highContrast: workspaceConfiguration.get<boolean>("highContrast"),
     };
   } // }}}
@@ -47,7 +44,6 @@ export default class Utils {
       configuration.darkWorkbench === "material" &&
       configuration.lightContrast === "medium" &&
       configuration.darkContrast === "medium" &&
-      configuration.diagnosticTextBackgroundOpacity === "0%" &&
       configuration.highContrast === false
     );
   } // }}}